Index Bug Fixes
Index bug fixes for the main hero image on iOS. Not fully resolved yet,
but fixed the gradient overlay that wasn't displaying properly.

Previously it was covering hte entire screen due to the "Cover" css
property. This caused it to not repeat correctly and overlay the image
like a good little overlay is supposed to do.